+++ cmd/clusterctl/Dockerfile
@@ -1,5 +1,5 @@
 # Build the cluster binary
-FROM golang:1.10.3 as builder
+FROM golang:1.12 as builder
 
 # Copy in the go src
 WORKDIR /go/src/sigs.k8s.io/cluster-api-provider-vsphere
