+++ docs/self-service/README.md
@@ -2,16 +2,17 @@
 
 The following is a quick how-to-use guide on using the cluster api on a vCenter infrastructure.  Before beginning, make sure you have the following requirements:
 
-1. vCenter 6.5 cluster
+1. vCenter 6.5+ cluster or ESXi 6.5+
     - You will need to gather some information about the cluster, described below.
-2. golang 1.10+
+    - [ESXi provisioning](./esx_provisioning.md) should be used for development purposes only.
+2. golang 1.12+
 3. a type 2 desktop hypervisor (Vmware Fusion/Workstation or VirtualBox)
 4. GOPATH environment set
 5. dep installed (https://github.com/golang/dep)
 6. kustomize V2 installed (https://github.com/kubernetes-sigs/kustomize/blob/master/docs/INSTALL.md)
 7. kubebuilder installed (https://book.kubebuilder.io/quick_start.html)
 
-**Be aware, the current repo supports deploying Kubernetes 1.11.x and above.  Older versions maybe supported in the future.**
+**Be aware, the current repo supports deploying Kubernetes 1.11.x and above.**
 
 ### Decide on whether to use minikube or kind as your bootstrapper
